requirements.txt
alembic==1.13.1
apispec==6.6.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==2.2.3
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
flask-smorest==0.42.3
Flask-SQLAlchemy==2.5.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.2
Mako==1.3.1
MarkupSafe==2.1.2
marshmallow==3.21.3
passlib==1.7.4
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
SQLAlchemy==1.4.49
webargs==8.4.0
Werkzeug==2.2.3
